==> WGCNA/consensus/2-consensus_net_const.Rmd <==
---
title: "Consensus block-wise network construction"
author: "Sofia Salazar"
date: "2023-10-18"
output: html_document
---

```{bash}
qlogin
module load r/4.0.2
R
```

Following this [tutorial](https://horvath.genetics.ucla.edu/html/CoexpressionNetwork/Rpackages/WGCNA/Tutorials/Consensus-NetworkConstruction-blockwise.pdf) for "An automatic block-wise network construction and module detection method for users who wish to analyze data sets too large to be analyzed all in one."

## Libraries & setup

```{r}
library(WGCNA)
options(stringsAsFactors = FALSE)
enableWGCNAThreads(15)
indir <- "/mnt/Citosina/amedina/ssalazar/meta/combined/coexpression/"
figdir <- "/mnt/Citosina/amedina/ssalazar/meta/combined/coexpression/figures/"
```

## Load data

```{r}
lnames = load(file = paste0(indir, "Consensus-dataInput.RData"))
lnames
```

```
[1] "multiExpr"   "Traits"      "nGenes"      "nSamples"    "setLabels"  
[6] "shortLabels" "exprSize"
```

```{r}
nSets = checkSets(multiExpr)$nSets # 2
```

## Analysis

1. Choose the soft-thresholding power

```{r}
# Choose a set of soft-thresholding powers
powers = c(seq(4,30,by=1), seq(4,30, by=2))
# Initialize a list to hold the results of scale-free analysis
powerTables = vector(mode = "list", length = nSets);
# Call the network topology analysis function for each set in turn
for (set in 1:nSets)
powerTables[[set]] = list(data = pickSoftThreshold(multiExpr[[set]]$data, powerVector=powers,
verbose = 2)[[2]]);
collectGarbage();
```

Plot results

```{r}
# Plot the results:
colors = c("black", "red")
# Will plot these columns of the returned scale free analysis tables
plotCols = c(2,5,6,7)
colNames = c("Scale Free Topology Model Fit", "Mean connectivity", "Median connectivity",
"Max connectivity");
# Get the minima and maxima of the plotted points
ylim = matrix(NA, nrow = 2, ncol = 4);
for (set in 1:nSets)
{
for (col in 1:length(plotCols))
{
ylim[1, col] = min(ylim[1, col], powerTables[[set]]$data[, plotCols[col]], na.rm = TRUE);
ylim[2, col] = max(ylim[2, col], powerTables[[set]]$data[, plotCols[col]], na.rm = TRUE);
}
}
# Plot the quantities in the chosen columns vs. the soft thresholding power
sizeGrWindow(8, 6)

pdf(file = paste0(figdir, "cons_scaleFreeAnalysis.pdf"), wi = 8, he = 6);
par(mfcol = c(2,2));
par(mar = c(4.2, 4.2 , 2.2, 0.5))
cex1 = 0.7;
for (col in 1:length(plotCols)) for (set in 1:nSets)
{
if (set==1)
{
plot(powerTables[[set]]$data[,1], -sign(powerTables[[set]]$data[,3])*powerTables[[set]]$data[,2],
xlab="Soft Threshold (power)",ylab=colNames[col],type="n", ylim = ylim[, col],
main = colNames[col]);
addGrid();
}
if (col==1)
{
text(powerTables[[set]]$data[,1], -sign(powerTables[[set]]$data[,3])*powerTables[[set]]$data[,2],
labels=powers,cex=cex1,col=colors[set]);
} else
text(powerTables[[set]]$data[,1], powerTables[[set]]$data[,plotCols[col]],
labels=powers,cex=cex1,col=colors[set]);
if (col==1)
{
legend("bottomright", legend = setLabels, col = colors, pch = 20) ;
} else
legend("topright", legend = setLabels, col = colors, pch = 20) ;
}
dev.off();
```

We choose 23 for both sets 

## Block-wise network construnction and consensus module detection


```{r}
bnet = blockwiseConsensusModules(
multiExpr, maxBlockSize = 15000, power = 23, minModuleSize = 30,
deepSplit = 2, TOMType = 'signed', saveConsensusTOMs = T, consensusTOMFilePattern = "/mnt/Citosina/amedina/ssalazaar/meta/combined/coexpression/tom/consensusTOM-block.%b.RData",
pamRespectsDendro = FALSE,
mergeCutHeight = 0.25, numericLabels = TRUE,
minKMEtoStay = 0, nThreads = 30, verbose = 5)
```

## Save

```{r}
save(bnet, file = paste0(indir, "Consensus-NetworkConstruction-auto.RData"))
```
